+++ plugins/packages/langgraph/src/kitaru_langgraph/__init__.py
@@ -1,0 +1,37 @@
+#  Copyright (c) ZenML GmbH 2026. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+"""LangGraph recording and replay adapter."""
+
+from .agent import KitaruGraphRunner
+from .capability import (
+    CapabilityOperation,
+    CapabilityTarget,
+    CapabilityTargetKind,
+    LangGraphAdapterError,
+    LangGraphCapabilityView,
+    LocalSubagentFactorySpec,
+    ToolPolicyError,
+    ToolPolicyMissError,
+    UnsupportedCapabilityError,
+    UnsupportedInvocationError,
+    UnsupportedWorkerInterruptError,
+)
+from .capture import CapturePolicy
+
+__all__ = [
+    "CapabilityOperation",
+    "CapabilityTarget",
+    "CapabilityTargetKind",
+    "CapturePolicy",
+    "KitaruGraphRunner",
+    "LangGraphAdapterError",
+    "LangGraphCapabilityView",
+    "LocalSubagentFactorySpec",
+    "ToolPolicyError",
+    "ToolPolicyMissError",
+    "UnsupportedCapabilityError",
+    "UnsupportedInvocationError",
+    "UnsupportedWorkerInterruptError",
+]